Provide clients a list of the users's features
Rather than clients compiling this themselves by checking for features _that client knows about_, let the server be the authority and provide a full list to the client.

This list could include features the client doesn't know about, so is useful for checking things like "is cape the only feature this user has?"

It also makes producing lists of features more accurate, such as the list of features "incognito" will disable shown in the Account Dashboard.